Shopping cart!

+++ estore-ui/src/app/services/cart.service.ts
@@ -1,11 +1,12 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders} from "@angular/common/http";
 import {Shoe} from "../ShoeInterface";
 import {Observable, of} from "rxjs";
-import {catchError, tap} from "rxjs/operators";
+import {catchError, map, tap} from "rxjs/operators";
+import {UserService} from "./user.service";
 
 @Injectable({
-  providedIn: 'root'
+	providedIn: 'root'
 })
 export class CartService {
 	httpOptions = {
@@ -13,40 +14,63 @@ export class CartService {
 	};
 	private cartURL = 'http://localhost:8080/cart';  // URL to web api
 
-	constructor(public http: HttpClient) {}
-	items: Shoe[] = [];
-
-	getItems() {
-		return this.items;
+	constructor(public http: HttpClient, public userService: UserService) {
 	}
 
-	addItem(item: Shoe): Observable<Shoe> {
-			this.items.push(item);
-			const url = `${this.cartURL}/`;
-			return this.http.post<Shoe>(url, item, this.httpOptions).pipe(
-				tap((shoe: Shoe) => console.log(`added cart w/ id=${shoe.id}`)),
-				catchError(this.handleError<Shoe>('addShoe'))
-			);
+	getItems(): Observable<Shoe[]> {
+		let user = this.userService.getUser();
+		if (user.username == "Anonymous") {
+			return of([]);
+		}
+
+		const url = `${this.cartURL}/${this.userService.getUser().username}`;
+		return this.http.get<Shoe[]>(url).pipe(
+			tap(_ => console.log(`Fetched shoes from cart`)),
+			map(obj => (obj as any).items),
+			catchError(this.handleError<Shoe[]>('getShoes'))
+		);
+	}
+
+	addItem(item: Shoe) {
+		let user = this.userService.getUser();
+		if (user.username == "Anonymous") {
+			return of([]);
+		}
+
+		const url = `${this.cartURL}/${this.userService.getUser().username}`;
+		return this.http.post<Shoe>(url, item, this.httpOptions).pipe(
+			catchError(this.handleError<Shoe>('addShoe'))
+		).subscribe(() => {
+			console.log("Added shoe to cart");
+		});
 	}
 
 	removeItem(item: Shoe) {
-		const index = this.items.indexOf(item);
-		if (index > -1) {
-			this.items.splice(index, 1);
+		let user = this.userService.getUser();
+		if (user.username == "Anonymous") {
+			return of([]);
 		}
-		const url = `${this.cartURL}/${item.id}`;
+
+		const url = `${this.cartURL}/${this.userService.getUser().username}`;
 		return this.http.delete<Shoe>(url, this.httpOptions).pipe(
 			tap(_ => console.log(`Removed shoe id=${item.id}`)),
 			catchError(this.handleError<Shoe>('Remove Shoe'))
 		);
 	}
 
-	getTotalCost(): number {
-		let totalCost = 0;
-		for (const item of this.items) {
-			totalCost += item.price;
+	getTotalCost(): Observable<number> {
+		let user = this.userService.getUser();
+		if (user.username == "Anonymous") {
+			return of(0);
 		}
-		return totalCost;
+
+		return this.getItems().pipe(
+			map((shoes: Shoe[]) => {
+				return shoes.reduce((prev, curr) => {
+					return prev + curr.price;
+				}, 0);
+			})
+		);
 	}
 
 	private handleError<T>(operation = 'operation', result?: T) {
